Cambiados test de la visualización de votaciones Cambiados los test de visualización de las votaciones al paquete visualizer ya que aunque comprueban nuevas funcionalidades añadidas en voting forman mas parte del testeo del correcto funcionamiento del visualizer y su template
